Refactored CommentThreadStore to support upcoming changes
Split into a base CommentThreadStore and StoryCommentThreadStore - if we're
 going to transition to more comment functionality being controlled by a
store passed through props, comments which aren't being displayed in the
context of a story will also need a store they can use.

Updated tests to account for StoryCommentThreadStore being a constructor
now, and for comment objects rather than ids being passed to it.